refactor: DTO 명시 및 spec 패키지명 code로 변경
